Update and rename seurat_preprocess_fun.R to seurat2shiny.R
I've reimplemented the original function by @MarcElosua with base R. No need to call the external packages dplyr, tibble and SummarizedExperiment. In fact, SummarizedExperiment was not used at all.

The metadata and expression objects are now saved as RDS files by default to be loaded in the ShinyApp. A list with the two objects is still returned but there's no need to assign it to a variable.

Finally, numeric columns in the metadata dataframe are transformed to factors by default, as we usually want to plot them as discrete values instead of continuous numeric values (in contrast to the expression values).